+++ lib/src/injector/injector.dart
@@ -6,4 +6,5 @@ typedef T DependencyInjectorBuilder<T>(Injector injector);
 abstract class Injector {
   void registerDependency<T>(DependencyInjectorBuilder<T> builder, {String dependencyName, bool isSingleton = false});
   T getDependency<T>({String dependencyName});
+  void reset();
 }